requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.8.3
PyMuPDF==1.23.8
python-docx==1.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.24.3
